backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
sqlalchemy==2.0.31
alembic==1.13.2
asyncpg==0.29.0
redis==5.0.7
celery==5.4.0
anthropic==0.31.0
httpx==0.27.0
pandas==2.2.2
numpy==1.26.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
cryptography==42.0.8
boto3==1.34.144
google-cloud-firestore==2.16.1
pysftp==0.2.9
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pymysql==1.1.1
pymongo==4.8.0
odata==0.4.1
zeep==4.2.1
xmltodict==0.13.0
aiofiles==23.2.1
tenacity==8.3.0
structlog==24.2.0
prometheus-fastapi-instrumentator==7.0.0
pytest==8.2.2
pytest-asyncio==0.23.7
